fix(openapi): use standard OpenAPI format with min/max for unsigned integers
Fixes #760

Previously, unsigned integer types (u8, u16, u32, u64, usize) used
non-standard format values like 'uint16', 'uint32', 'uint64' in the
generated OpenAPI schema. These formats are not part of the OpenAPI
specification and can cause issues with code generators and validators.

This change:
- Uses standard 'int32' format for u8 and u16 (they fit in 32 bits)
- Uses standard 'int64' format for u32, u64, and usize
- Adds 'minimum: 0' constraint to indicate unsigned values
- Adds 'maximum' constraint with the type's max value (e.g., 255 for u8)
- Applies the same fix to NonZero<uX> types with exclusive_minimum: true

The generated schema now produces valid OpenAPI that accurately
describes the range constraints of unsigned integers.

+++ poem-openapi/tests/api.rs
@@ -469,7 +469,7 @@ async fn bad_request_handler() {
     resp.assert_status(StatusCode::BAD_REQUEST);
     resp.assert_content_type("text/plain; charset=utf-8");
     resp.assert_text(
-        r#"!!! failed to parse parameter `code`: Type "integer(uint16)" expects an input value."#,
+        r#"!!! failed to parse parameter `code`: Type "integer(int32)" expects an input value."#,
     )
     .await;
 }
